feat: extract tree functionality into feature module with Zustand store
