+++ src/features/tree/index.ts
@@ -1,0 +1,16 @@
+export { useTreeOperations } from "./useTreeOperations";
+export { useTreeStore } from "./treeStore";
+export type { TreeState } from "./treeStore";
+
+// Components
+export { Tree } from "./Tree";
+
+// Utilities
+export {
+  createNodesFromJSON,
+  createNodeFromValue,
+  getValueAtPointer,
+} from "./treeUtils";
+
+// Re-export for convenience
+export type { TreeNodeProps } from "@/shared/types";
